Simplify the user churn code some
The old code was somewhat complicated in an attempt to reliably
produce user churn day over day. Due to a bug with 'conjur user
retire' (#106344362) it was also always broken. The new code is more
reliable and works around the above bug but will not generate as
consistent of results day-to-day due to increased randomness. However
over a long run it should prove more stable.