Save API keys from populate.sh.
Because the context files have different formats in the old and new dsl,
the `script execute` commands save their contexts to `conjur.json`, while
the `policy2 load` commands save theirs to `api-keys.json`.  This reflects
the fact that the new dsl simply writes the api keys for roles it creates
as a hash of `roleid => api_key`, whereas the old dsl writes additional information such as certs and an appliance url.